[tile] Flatten BOOM Tile organization
